file structure overhaul and setprecision() fix

+++ Student_Code/Student.hpp
@@ -11,6 +11,7 @@
 
 #include <stdio.h>
 #include <sstream>
+#include <iomanip>
 #include "StudentInterface.h"
 
 class Student : public StudentInterface {
@@ -19,7 +20,7 @@ protected:
     string m_Name;
     string m_Address;
     string m_Phone;
-    int m_GPA;
+    float m_GPA;
     
 public:
     Student(int num_id, string name, string address, string phone);
